play mode: decouple Play from drive-mode preset
Play is now a generic lifecycle (play-mode system: start/stop +
scene events), and drive-mode is one of potentially many subscribers.
Sets the stage for traffic animation as a sibling feature that
listens for the same events without depending on drive-mode or Rapier.

- new src/aframe-components/play-mode.js (system: isPlaying + events)
- drive-mode scene component moved from viewer-mode.js into
  play-mode-vehicle.js, attached via <a-scene drive-mode>
- viewer-mode loses the 'drive' preset and its drive-only paths
- Play button just calls play-mode.start(); Stop just reopens the
  inspector (store handles play-mode.stop() on open)
- play-mode-vehicle emits 'vehicle-built' so PlayModeControls can
  drop its rAF polling
- fixes Stop-before-Rapier-loaded race with an activation token
