+++ index.html
@@ -82,7 +82,8 @@
     reflection device-orientation-permission-ui="enabled: false"
     webxr="requiredFeatures:hit-test,local-floor;referenceSpaceType:local-floor;"
     xr-mode-ui="XRMode: ar; enterARButton: #viewer-mode-ar-play-button;"
-    css2d-renderer scene-timer="autoStart: false; format: mm:ss:ff">
+    css2d-renderer scene-timer="autoStart: false; format: mm:ss:ff"
+    drive-mode>
     <a-assets>
       <!-- TODO: Add this to repo documentation  -->
       <!-- you can specify a custom asset path using below syntax  -->
