+++ src/index.js
@@ -39,6 +39,7 @@ require('./aframe-components/measure-line.js');
 require('./aframe-components/css2d-renderer.js');
 require('./aframe-components/google-maps-aerial.js');
 require('./aframe-components/viewer-mode.js');
+require('./aframe-components/play-mode.js');
 require('./aframe-components/play-mode-vehicle.js');
 require('./aframe-components/scene-timer.js');
 require('./aframe-components/geojson.js');
